services/map-processor/requirements.txt
flask
google-cloud-storage
numpy
functions-framework
